checkpoint: renames "/app.ts" to "/external.ts" in "/vue"
