+++ src/library/vue/external.ts
